+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.87 ] - 2026/05/15
+
+  * Added model `NowSelectionUpdated` - device NowSelectionUpdated event configuration object.
+
 ###### [ 1.0.86 ] - 2025/11/03
 
   * Added method `SoundTouchClient.SetProductHdmiAssignmentControls` to set the current product HDMI assignment controls configuration of the device.